[Passed] 두 수의 거듭제곱 (2ms, 0MB)

+++ trail2/두 수의 거듭제곱/two-numbers-of-squares.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int a, b;
 
 // ex) 3^13 = 3^8 * 3^4 * 3^1 (1101)
-unsigned long long power(int base, int exp) {
+unsigned long long power(unsigned long long base, int exp) {
     unsigned long long res = 1;
     while (exp > 0) {
         // & 1 비트연산으로 홀수(첫 비트 1)인 경우 추출
